Add docker_up.sh, docker_down.sh, and services_ping.sh - Documentation Update (#50)
* Add Mac Os Steps for VS Code

Added additional steps to follow in order to allow the command 'code .' in terminal.

* Add docker_up.sh, docker_down.sh, and services_ping.sh

Added Bash Scripts to better align with their PowerShell equivalents: docker_up.sh which is a bash version of docker_up.ps1 script, taking in -r and -b commands. docker_down.sh which is a bash version of docker_down.ps1.  services_ping.sh which is a bash version of services_ping.ps1.

* Add Running Under Docker Bash Documentation

Updated documentation to show how to run under docker using Terminal and Bash on Mac or Linux.

* Fix Eve & Carter profile link

Fixed broken link to Eve and Carter.